Move _ValueClass selection out of core.py
