+++ prometheus_client/core.py
@@ -3,7 +3,6 @@
 from __future__ import unicode_literals
 
 import math
-import os
 import sys
 import time
 import types
@@ -14,47 +13,14 @@ from .registry import REGISTRY
 from .metrics import *  # noqa
 from .metrics import _METRIC_LABEL_NAME_RE, _METRIC_NAME_RE, _RESERVED_METRIC_LABEL_NAME_RE
 from .context_managers import _ExceptionCounter, _InprogressTracker, _Timer
+from .value import get_value_class
 
 if sys.version_info > (3,):
     unicode = str
 
 _INF = float("inf")
 _MINUS_INF = float("-inf")
-
-
-
-class _MutexValue(object):
-    '''A float protected by a mutex.'''
-
-    _multiprocess = False
-
-    def __init__(self, typ, metric_name, name, labelnames, labelvalues, **kwargs):
-        self._value = 0.0
-        self._lock = Lock()
-
-    def inc(self, amount):
-        with self._lock:
-            self._value += amount
-
-    def set(self, value):
-        with self._lock:
-            self._value = value
-
-    def get(self):
-        with self._lock:
-            return self._value
-
-
-# Should we enable multi-process mode?
-# This needs to be chosen before the first metric is constructed,
-# and as that may be in some arbitrary library the user/admin has
-# no control over we use an environment variable.
-if 'prometheus_multiproc_dir' in os.environ:
-    from .multiprocess import _MultiProcessValue
-
-    _ValueClass = _MultiProcessValue()
-else:
-    _ValueClass = _MutexValue
+_ValueClass = get_value_class()
 
 
 class _LabelWrapper(object):
