+++ prometheus_client/value.py
@@ -1,0 +1,41 @@
+import os
+from threading import Lock
+
+__all__ = (
+    '_MutexValue',
+    'get_value_class',
+)
+
+
+class _MutexValue(object):
+    '''A float protected by a mutex.'''
+
+    _multiprocess = False
+
+    def __init__(self, typ, metric_name, name, labelnames, labelvalues, **kwargs):
+        self._value = 0.0
+        self._lock = Lock()
+
+    def inc(self, amount):
+        with self._lock:
+            self._value += amount
+
+    def set(self, value):
+        with self._lock:
+            self._value = value
+
+    def get(self):
+        with self._lock:
+            return self._value
+
+
+def get_value_class():
+    # Should we enable multi-process mode?
+    # This needs to be chosen before the first metric is constructed,
+    # and as that may be in some arbitrary library the user/admin has
+    # no control over we use an environment variable.
+    if 'prometheus_multiproc_dir' in os.environ:
+        from .multiprocess import _MultiProcessValue
+        return _MultiProcessValue()
+    else:
+        return _MutexValue
